+++ NEWS.md
@@ -4,6 +4,7 @@
   - `download_yq()` now supports the masOS (`darwin`).
   - Fixed URL to Windows version (`.exe` extension).
   - Better OS resolution implemented in `.get_os()`.
+- Fixed issues with `yq` tool's binary download on Windows.
 
 # scdrake 1.0.0